% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_syns.R
\name{bind_synonyms_}
\alias{bind_synonyms}
\alias{bind_synonyms_}
\title{Bind a column of synonyms to a dataframe of words}
\usage{
bind_synonyms_(tbl, words, pos = NULL, unnest = FALSE, min_rel_level = 1,
  kill_ALL_conns_on_exit = TRUE)

bind_synonyms(tbl, words, pos = NULL, unnest = FALSE, min_rel_level = 1,
  kill_ALL_conns_on_exit = TRUE)
}
\arguments{
\item{tbl}{dataframe to bind synonyms to}

\item{words}{A vector of UNIQUE words to return synonyms for}

\item{pos}{Optional vector of part of speech tags corresponding to \code{words}. Returns synonyms for all parts of speech available if set to NULL.  Will ignore elements passed as \code{NA} and will return synonyms for all parts of speech}

\item{unnest}{Should list column of synonym vectors be unnested with \code{tidyr::unnest}}

\item{min_rel_level}{Getting synonyms from thesaurus.com will return a relevance rank (1 being most relevant); this param sets the minumum rank to return}

\item{kill_ALL_conns_on_exit}{If a url can't be found, a connection will linger creating a warning message about closing unused connections.  This parameter will close these connections w/o warning; set to FALSE if you are using other connections that you dont want closed.}
}
\value{
A list of equal length to \code{words}.  Each list is a character vector of the synonyms found. If no synomyms found then NA will be returned for that word part of speech pair
}
\description{
Bind a column of synonyms to a dataframe of words
}
\examples{
my_df <- data.frame(my_words=c("test","test"),
                   my_pos=c("noun","verb"),
                   stringsAsFactors = FALSE)

bind_synonyms(my_df, my_words, my_pos)
bind_synonyms_(my_df, "my_words", "my_pos")
}

