% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_com_syns.R
\name{get_syns}
\alias{get_syns}
\title{Get synonyms for UNIQUE word part of speech pairs}
\usage{
get_syns(words, pos = NULL, min_rel_level = 1, source = "thesaurus.com",
  kill_ALL_conns_on_exit = TRUE)
}
\arguments{
\item{words}{A vector of UNIQUE words to return synonyms for}

\item{pos}{Optional vector of part of speech tags corresponding to \code{words}. Returns synonyms for all parts of speech available if set to NULL.  Will ignore elements passed as \code{NA} and will return synonyms for all parts of speech}

\item{min_rel_level}{Getting synonyms from thesaurus.com will return a relevance rank (1 being most relevant); this param sets the minumum rank to return}

\item{source}{literally does nothing right now. plan to add more sources for synonyms in future}

\item{kill_ALL_conns_on_exit}{If a url can't be found, a connection will linger creating a warning message about closing unused connections.  This parameter will close these connections w/o warning; set to FALSE if you are using other connections that you dont want closed.}
}
\value{
A list of equal length to \code{words}.  Each list is a character vector of the synonyms found. If no synomyms found then NA will be returned for that word part of speech pair
}
\description{
Get synonyms for UNIQUE word part of speech pairs
}
\examples{
get_syns(c("test","test"),c("noun","verb"))
}

