% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_com_syns_helper.R
\name{get_thesaurus.com_syns_df}
\alias{get_thesaurus.com_syns_df}
\title{Get synonyms from thesaurus.com with their relevance scores}
\usage{
get_thesaurus.com_syns_df(word_slash_pos)
}
\arguments{
\item{word_slash_pos}{word and part of speech passed as <word/pos> if not providing pos no slash is needed after word}
}
\value{
A tibble with 2 columns: synonym & relevance
}
\description{
Get synonyms from thesaurus.com with their relevance scores
}
\examples{
get_thesaurus.com_syns_df("fly")
get_thesaurus.com_syns_df("fly/verb")
}

